/fixes/ race condition
Well this is not /really/ a fix, more like a workaround. There is a
race condition when a directory gets added to a tracked directory and
files are inserted inside it before a handler was attached to it. For
example, lets asume we are tracking folder A. Then 10 files are
inserted into A, then a directory B and finally 10 more files into
B. When files are getting inserted into B while we are still handling
the file events from A, there wont be any callback provied for those,
since B does not have a handler yet. Or even worse, if there is a
folder C inserted into B befor a handler is attached on B we would
even miss the entire folder C and all its subdirectories. To
(somewhat) fix this two loops will now iterate over new folders and
add handler for all subdirectories (recursivly) plus call the callback
function for each file as if it was added. The ugly part here is that
we can get two :file-created events for a single file...