adds EVENT-LOOP-BUSY-P and HOOK-BUSY-P
Adds two new slots which are used to check if the watcher is
busy. EVENT-LOOP-BUSY-P is t while CALLBACK is called (when a event
gets processed inside the event-loop). HOOK-BUSY-P is t while a Hook
is being run by the hook-thread.

Also adds alive-p check to BUSY-P, because if the watcher was
initialized and started, alive-p gets set _after_ all initial
directories where added. Which basically means that alive-p is nil
while the watcher is initializing. There is also a check if the thread
is alive, since if the watcher was stopped, busy-p should return NIL.