fixes wildcards bug with pathnames
When a tracked file or folder contained a * ? or [ character a error
was thrown. This is now fixed by wrapping the pathnames with the
escape-wildcards function which escapes the characters. This is not a
perfect solution and might still throw a error on other
implementations besides sbcl. But its at least fixed on sbcl.